refactor!: move `vcs` module into `repo` module
The `vcs` module is only going to be used by the `repo` module. Thus, it makes
more sense to move it into the `repo` module itself.
